+++ Tests/Unit/DependencyInjection/CmfSeoExtensionTest.php
@@ -4,6 +4,7 @@ namespace Symfony\Cmf\SeoBundle\Tests\Unit\DependencyInjection;
 
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionTestCase;
 use Symfony\Cmf\Bundle\SeoBundle\DependencyInjection\CmfSeoExtension;
+use Symfony\Component\DependencyInjection\Definition;
 
 class CmfSeoExtensionTest extends AbstractExtensionTestCase
 {
@@ -38,6 +39,7 @@ class CmfSeoExtensionTest extends AbstractExtensionTestCase
             array(
                 'CmfRoutingBundle' => true,
                 'SonataDoctrinePHPCRAdminBundle' => true,
+                'DoctrinePHPCRBundle' => true,
             )
         );
         $this->load(array(
@@ -75,6 +77,7 @@ class CmfSeoExtensionTest extends AbstractExtensionTestCase
             array(
                 'CmfRoutingBundle' => true,
                 'SonataDoctrineORMBundle' => true,
+                'DoctrinePHPCRBundle' => true,
             )
         );
 
@@ -88,5 +91,51 @@ class CmfSeoExtensionTest extends AbstractExtensionTestCase
         ));
 
         $this->assertContainerBuilderHasService('cmf_seo.admin_extension', 'Symfony\Cmf\Bundle\SeoBundle\Admin\Extension\SeoContentAdminExtension');
+    }
+
+    public function testAlternateLocaleWithPhpcr()
+    {
+        $this->container->setParameter(
+            'kernel.bundles',
+            array(
+                'DoctrinePHPCRBundle' => true,
+                'CmfRoutingBundle' => true,
+            )
+        );
+        $this->load(array(
+            'persistence'   => array(
+                'phpcr' => true,
+            ),
+            'alternate_locale' => array(
+                'enabled' => true
+            ),
+        ));
+
+        $this->assertContainerBuilderHasService(
+            'cmf_seo.alternate_locale.provider_phpcr',
+            'Symfony\Cmf\Bundle\SeoBundle\Doctrine\Phpcr\AlternateLocaleProvider'
+        );
+
+        $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
+            'cmf_seo.event_listener.seo_content',
+            'setAlternateLocaleProvider',
+            array($this->container->getDefinition('cmf_seo.alternate_locale.provider_phpcr'))
+        );
+    }
+
+    public function testAlternateLocaleWithCustomProvider()
+    {
+        $this->container->setDefinition('some_alternate_locale_provider', new Definition());
+        $this->load(array(
+            'alternate_locale' => array(
+                'provider_id' => 'some_alternate_locale_provider'
+            ),
+        ));
+
+        $this->assertContainerBuilderHasServiceDefinitionWithMethodCall(
+            'cmf_seo.event_listener.seo_content',
+            'setAlternateLocaleProvider',
+            array($this->container->getDefinition('some_alternate_locale_provider'))
+        );
     }
 }