+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 1.1.0-RC1
 ---------
 
+* **2014-02-08**: Implement alternate locale support and its configuration
 * **2014-06-06**: Updated to PSR-4 autoloading
 
 1.0.0